requirements.txt
numpy<2
pydub
essentia-tensorflow
torch
torchaudio
soundfile
timm==0.4.5